Finish Develop Customer Backend

+++ btr.domain/SalesContext/FakturAgg/FakturModel.cs
@@ -16,8 +16,8 @@ namespace btr.domain.SalesContext.FakturAgg
     
         public string CustomerId { get; set; }
         public string CustomerName { get; set; }
-        public double Plafond { get; set; }
-        public double CreditBalance { get; set; }
+        public decimal Plafond { get; set; }
+        public decimal CreditBalance { get; set; }
     
         public string WarehouseId { get; set; }
         public string WarehouseName { get; set; }
